Split parse/nodes module into nodes/ package with facade
