--- a816/parse/nodes.py
@@ -1,1067 +1,0 @@
-import logging
-import re
-import struct
-
-from a816.cpu.cpu_65c816 import (
-    NoOpcodeForOperandSize,
-    guess_value_size,
-    snes_opcode_table,
-)
-from a816.cpu.mapping import Address
-from a816.cpu.types import AddressingMode, ValueSize
-from a816.diagnostics.suggest import did_you_mean_hint as _did_you_mean_hint
-from a816.error_codes import E_SYMBOL_NOT_DEFINED as _E_SYMBOL_NOT_DEFINED
-from a816.exceptions import A816Error, ExternalExpressionReference, ExternalSymbolReference, SymbolNotDefined
-from a816.object_file import Region
-from a816.parse.ast.expression import eval_expression, eval_expression_str
-from a816.parse.ast.nodes import BlockAstNode, ExpressionAstNode
-from a816.parse.tokens import Token
-from a816.pool import Allocation
-from a816.protocols import NodeProtocol, OpcodeProtocol, ValueNodeProtocol
-from a816.symbols import Resolver, Scope
-from script import Table
-
-logger = logging.getLogger("a816.nodes")
-
-
-class ValueNode(ValueNodeProtocol):
-    def __init__(self, value: str) -> None:
-        self.value = value
-
-    def get_value(self) -> int:
-        return int(self.value, 16)
-
-    def get_value_string_len(self) -> int:
-        value_length = len(self.value)
-        return value_length
-
-    def __str__(self) -> str:
-        return f"ValueNode({self.value})"
-
-
-class ExpressionNode(ValueNodeProtocol):
-    def __init__(self, expression: ExpressionAstNode, resolver: Resolver, file_info: Token) -> None:
-        self.expression = expression
-        self.resolver = resolver
-        self.file_info = file_info
-
-    def _compute_local_label_renames(self) -> tuple[dict[str, str], bool]:
-        """Return (rename map, touches_any_label). Nested-scope label refs get mangled."""
-        from a816.parse.tokens import TokenType
-
-        rename: dict[str, str] = {}
-        touches_label = False
-        for t in self.expression.tokens:
-            tok = getattr(t, "token", None)
-            if tok is None or tok.type != TokenType.IDENTIFIER:
-                continue
-            owner = self.resolver.current_scope.find_label_scope(tok.value)
-            if owner is None:
-                continue
-            touches_label = True
-            if owner is self.resolver.scopes[0]:
-                continue  # root labels keep their plain name; only nested-scope refs need mangling
-            scope_idx = self.resolver.scopes.index(owner)
-            rename[tok.value] = f"__sc{scope_idx}__{tok.value}"
-        return rename, touches_label
-
-    def _record_local_label_relocation(self) -> None:
-        from a816.parse.ast.expression import _inline_aliases, reconstruct_expression
-
-        rename, touches_label = self._compute_local_label_renames()
-        if not touches_label:
-            return
-        expr_str = _inline_aliases(reconstruct_expression(self.expression), self.resolver)
-        for short, mangled in rename.items():
-            expr_str = re.sub(rf"\b{re.escape(short)}\b", mangled, expr_str)
-        self._deferred_expression = expr_str
-        self._local_label_renames = rename
-
-    def get_value(self) -> int | str:  # type:ignore
-        try:
-            value = eval_expression(self.expression, self.resolver)
-            if self.resolver.context.is_object_mode and isinstance(value, int):
-                # Module-local label refs: record the original expression so the
-                # linker can re-evaluate against the module's final placement.
-                self._record_local_label_relocation()
-            return value
-        except ExternalExpressionReference as e:
-            if self.resolver.context.is_object_mode:
-                self._deferred_expression = e.expression_str
-                self._external_symbols = e.external_symbols
-                return 0
-            raise NodeError(f"Expression contains external symbols: {e.expression_str}", self.file_info) from e
-        except ExternalSymbolReference as e:
-            if self.resolver.context.is_object_mode:
-                self._deferred_expression = self.expression.to_representation()[0]
-                return 0
-            raise NodeError(f"{e} ({self}) is not defined in the current scope.", self.file_info) from e
-        except SymbolNotDefined as e:
-            raise NodeError(
-                f"`{e}` is not defined in the current scope",
-                self.file_info,
-                code=str(_E_SYMBOL_NOT_DEFINED),
-                hint=_did_you_mean_hint(str(e), self.resolver.current_scope),
-            ) from e
-
-    def get_value_string_len(self) -> int:
-        value = self.get_value()
-        if not isinstance(value, int):
-            raise TypeError(f"Expected int, got {type(value).__name__}")
-        return len(hex(value)) - 2
-
-    def __str__(self) -> str:
-        return f"{self.__class__.__name__}({self.expression.to_representation()[0]})"
-
-
-class LabelNode(NodeProtocol):
-    def __init__(self, symbol_name: str, resolver: Resolver) -> None:
-        self.symbol_name = symbol_name
-        self.resolver = resolver
-
-    def emit(self, current_addr: Address) -> bytes:
-        return b""
-
-    def pc_after(self, current_pc: Address) -> Address:
-        self.resolver.current_scope.add_label(self.symbol_name, current_pc)
-        return current_pc
-
-    def __repr__(self) -> str:
-        return self.__str__()
-
-    def __str__(self) -> str:
-        return f"LabelNode({self.symbol_name})"
-
-
-class LabelDeclNode(NodeProtocol):
-    """`.label NAME = ADDR` — register NAME as a label at constant address ADDR.
-
-    Position counter is untouched. The address binds via
-    `scope.absolute_labels` (separate from `scope.labels` so the linker
-    doesn't shift it by the module relocation delta) and lands in `.adbg`
-    as `SymbolKind.LABEL`, so `lookup_label(addr)` resolves the name.
-    The RHS must evaluate to an int at this resolution pass; external
-    references are not supported (use `.extern` for that).
-    """
-
-    def __init__(
-        self,
-        symbol_name: str,
-        expression: ExpressionAstNode,
-        resolver: Resolver,
-        file_info: Token,
-    ) -> None:
-        self.symbol_name = symbol_name
-        self.expression = expression
-        self.resolver = resolver
-        self.file_info = file_info
-
-    def emit(self, current_addr: Address) -> bytes:
-        return b""
-
-    def pc_after(self, current_pc: Address) -> Address:
-        try:
-            value = eval_expression(self.expression, self.resolver)
-        except (ExternalExpressionReference, ExternalSymbolReference) as e:
-            ref = e.symbol_name if isinstance(e, ExternalSymbolReference) else e.expression_str
-            raise NodeError(
-                f".label {self.symbol_name}: address must be a constant expression (got external reference '{ref}')",
-                self.file_info,
-            ) from e
-        if not isinstance(value, int):
-            raise NodeError(
-                f".label {self.symbol_name}: address must evaluate to an int, got {type(value).__name__}",
-                self.file_info,
-            )
-        # The value is an absolute address the user supplied — not the
-        # current PC. Record it under `absolute_labels` (separate from
-        # `labels`) so the linker doesn't add the module's relocation delta
-        # to it, and write it directly into `symbols` so resolution at
-        # call sites returns the int. Skipping `add_symbol` avoids the
-        # "Symbol already defined" warning on the second resolve pass.
-        scope = self.resolver.current_scope
-        scope.absolute_labels[self.symbol_name] = value
-        scope.symbols[self.symbol_name] = value
-        return current_pc
-
-    def __str__(self) -> str:
-        return f"LabelDeclNode({self.symbol_name}, {self.expression})"
-
-
-class SymbolNode(NodeProtocol):
-    def __init__(
-        self,
-        symbol_name: str,
-        expression: ExpressionAstNode | BlockAstNode,
-        resolver: Resolver,
-    ) -> None:
-        self.symbol_name = symbol_name
-        self.expression = expression
-        self.resolver = resolver
-
-    def emit(self, current_addr: Address) -> bytes:
-        return b""
-
-    def _register_alias(self, expr_str: str) -> None:
-        self.resolver.current_scope.add_external_alias(self.symbol_name, expr_str)
-        object_writer = self.resolver.context.object_writer
-        if object_writer is not None:
-            object_writer.add_alias(self.symbol_name, expr_str)
-
-    def pc_after(self, current_pc: Address) -> Address:
-        # SymbolNode emits no bytes; current_pc is returned unchanged. The
-        # method exists to register the symbol's value (or alias) at resolution
-        # time. Protocol contract requires returning the PC.
-        assert isinstance(self.expression, ExpressionAstNode)
-        try:
-            value = eval_expression(self.expression, self.resolver)
-        except (ExternalExpressionReference, ExternalSymbolReference) as e:
-            if not self.resolver.context.is_object_mode:
-                raise NodeError(
-                    f"{self.symbol_name} = {self.expression.to_canonical()}: "
-                    f"external symbols only allowed in object compilation mode.",
-                    self.expression.file_info if hasattr(self.expression, "file_info") else current_pc,  # type: ignore[arg-type]
-                ) from e
-            self._register_alias(e.symbol_name if isinstance(e, ExternalSymbolReference) else e.expression_str)
-        else:
-            if self.resolver.context.is_object_mode and self._references_local_label():
-                # RHS hits a module-local CODE label. Register an alias so refs
-                # go through the relocation pipeline; baked value is
-                # module-base-relative.
-                from a816.parse.ast.expression import _inline_aliases, reconstruct_expression
-
-                self._register_alias(_inline_aliases(reconstruct_expression(self.expression), self.resolver))
-            else:
-                self.resolver.current_scope.add_symbol(self.symbol_name, value)
-        return current_pc
-
-    def _references_local_label(self) -> bool:
-        from a816.parse.tokens import TokenType
-
-        if not isinstance(self.expression, ExpressionAstNode):
-            return False
-        for term in self.expression.tokens:
-            tok = getattr(term, "token", None)
-            if tok is None or tok.type != TokenType.IDENTIFIER:
-                continue
-            if self.resolver.current_scope.find_label_scope(tok.value) is not None:
-                return True
-        return False
-
-    def __str__(self) -> str:
-        return f"SymbolNode({self.symbol_name}, {self.expression})"
-
-
-class ExternNode(NodeProtocol):
-    def __init__(self, symbol_name: str, resolver: Resolver) -> None:
-        self.symbol_name = symbol_name
-        self.resolver = resolver
-
-    def emit(self, current_addr: Address) -> bytes:
-        return b""
-
-    def pc_after(self, current_pc: Address) -> Address:
-        # Mark symbol as external in the current scope
-        # Late import: intentional to avoid circular dependency with object_file module
-        from a816.object_file import SymbolSection, SymbolType
-
-        # Add external symbol to the resolver's scope
-        self.resolver.current_scope.add_external_symbol(self.symbol_name)
-
-        # Add external symbol to the object writer if we're in object compilation mode
-        object_writer = self.resolver.context.object_writer
-        if self.resolver.context.is_object_mode and object_writer is not None:
-            object_writer.add_symbol(self.symbol_name, 0, SymbolType.EXTERNAL, SymbolSection.CODE)
-
-        return current_pc
-
-    def __str__(self) -> str:
-        return f"ExternNode({self.symbol_name})"
-
-
-class RegisterSizeNode(NodeProtocol):
-    """Node for register size directives (.a8, .a16, .i8, .i16)"""
-
-    def __init__(self, register: str, size: int, resolver: Resolver) -> None:
-        self.register = register  # "a" for accumulator, "i" for index
-        self.size = size  # 8 or 16
-        self.resolver = resolver
-
-    def emit(self, current_addr: Address) -> bytes:
-        # Update resolver state during emission
-        if self.register == "a":
-            self.resolver.a_size = self.size
-        else:
-            self.resolver.i_size = self.size
-        return b""
-
-    def pc_after(self, current_pc: Address) -> Address:
-        return current_pc
-
-    def __str__(self) -> str:
-        return f"RegisterSizeNode({self.register}{self.size})"
-
-
-class BinaryNode(NodeProtocol):
-    def __init__(self, path: str, resolver: Resolver) -> None:
-        with open(path, "rb") as binary_file:
-            self.binary_content = binary_file.read()
-        self.file_path = path
-        self.symbol_base = path.replace("/", "_").replace(".", "_")
-        self.resolver = resolver
-
-    def emit(self, current_addr: Address) -> bytes:
-        return self.binary_content
-
-    def pc_after(self, current_pc: Address) -> Address:
-        retval = current_pc + len(self.binary_content)
-        self.resolver.current_scope.add_label(self.symbol_base, current_pc)
-        self.resolver.current_scope.add_symbol(self.symbol_base + "__size", len(self.binary_content))
-        return retval
-
-
-class LinkedModuleNode(NodeProtocol):
-    """Emits a compiled module's regions and binds its symbols.
-
-    Each region has a compile-time `base_address` (from `*=`) and a code
-    blob. If the module is `relocatable` (no `*=` was present), region 0
-    gets shifted by `delta = import_pc - regions[0].base_address` and
-    every CODE symbol moves with it. Otherwise regions land at their
-    declared absolute addresses, ignoring the import site.
-    """
-
-    def __init__(
-        self,
-        module_name: str,
-        regions: list[Region],
-        symbols: list[tuple[str, int, int, int]],  # (name, address, type, section)
-        resolver: Resolver,
-        relocatable: bool = True,
-    ) -> None:
-        self.module_name = module_name
-        self.regions = regions
-        self.symbols = symbols
-        self.resolver = resolver
-        self.relocatable = relocatable
-        # Set by Program._mark_import_winners: True for every duplicate
-        # `.import` of the same module except the last occurrence in
-        # program order. Losers bind symbols (so the loser's source can
-        # still reference them) but do NOT advance the PC and do NOT
-        # emit bytes — the winner is the canonical placement.
-        self.is_loser: bool = False
-        self._delta = 0
-        # Cache placed regions for emit_blocks; refreshed on every pc_after.
-        self._placed: list[tuple[int, bytes]] = []
-
-    def emit(self, current_addr: Address) -> bytes:
-        # Single-region modules still flow through the legacy single-bytes
-        # path used by writers that don't know about emit_blocks.
-        del current_addr
-        placed = self._compute_placement()
-        if not placed:
-            return b""
-        if len(placed) == 1:
-            return placed[0][1]
-        # Multi-region modules must go through emit_blocks; returning the
-        # concatenation here would silently corrupt the output.
-        return b""
-
-    def emit_blocks(self, current_addr: Address) -> list[tuple[int, bytes]]:
-        del current_addr
-        return self._compute_placement()
-
-    def pc_after(self, current_pc: Address) -> Address:
-        self._local_map: dict[str, int] = {}
-        self._compute_delta_and_base(current_pc)
-        self._bind_module_symbols()
-        return self._advance_pc(current_pc)
-
-    def _compute_delta_and_base(self, current_pc: Address) -> None:
-        if self.regions:
-            self._delta = current_pc.logical_value - self.regions[0].base_address if self.relocatable else 0
-            # Shifted base of region 0 — used for the .sym/.adbg producer
-            # to report where the module actually landed.
-            self.base_address = self.regions[0].base_address + self._delta
-        else:
-            # Symbol-only module (e.g. a stubs file with only `.label`
-            # declarations and no emitted code). No regions means no delta
-            # and no PC advance, but symbols still need binding.
-            self._delta = 0
-            self.base_address = current_pc.logical_value
-
-    def _bind_module_symbols(self) -> None:
-        from a816.object_file import SymbolType
-
-        scope = self.resolver.current_scope
-        for name, address, sym_type, section in self.symbols:
-            final = self._resolve_symbol_address(address, section)
-            if sym_type == SymbolType.GLOBAL.value:
-                self._bind_global(scope, name, final, section)
-            elif sym_type == SymbolType.LOCAL.value:
-                self._local_map[name] = final
-
-    @staticmethod
-    def _bind_global(scope: Scope, name: str, final: int, section: int) -> None:
-        from a816.object_file import SymbolSection
-
-        scope.symbols[name] = final
-        if section == SymbolSection.CODE.value:
-            scope.labels[name] = final
-        elif section == SymbolSection.ABS_LABEL.value:
-            # `.label`-declared in the imported module — surface it as an
-            # absolute label in the importer's scope so it lands in the
-            # merged `.adbg` as SymbolKind.LABEL.
-            scope.absolute_labels[name] = final
-
-    def _advance_pc(self, current_pc: Address) -> Address:
-        # Loser duplicates publish symbols (winner overwrites later via
-        # last-pass) but must not consume PC space — otherwise inline
-        # source surrounding the loser .import shifts forward by the
-        # module's size and lands on top of unrelated ROM.
-        # Symbol-only modules (no regions) don't advance PC either.
-        # Pinned modules (any explicit `*=`) land at their declared
-        # absolute base addresses; the importer's PC stays where it was
-        # because the module does not occupy linear space at the import
-        # site. Only relocatable single-region modules advance the
-        # importer's PC by their first-region size.
-        if self.is_loser or not self.regions or not self.relocatable:
-            return current_pc
-        first = self.regions[0]
-        first_end = first.base_address + self._delta + len(first.code)
-        return self.resolver.get_bus().get_address(first_end)
-
-    def _resolve_symbol_address(self, address: int, section: int) -> int:
-        from a816.object_file import SymbolSection
-
-        if section != SymbolSection.CODE.value:
-            return address
-        return address + self._delta
-
-    def _compute_placement(self) -> list[tuple[int, bytes]]:
-        # Re-evaluate every call: cross-module symbols may have been bound
-        # by other LinkedModuleNodes after this one's pc_after ran. Doing
-        # the eval lazily at emit time avoids spurious "Failed to
-        # evaluate" warnings during the first resolve_labels pass.
-        placed: list[tuple[int, bytes]] = []
-        for region in self.regions:
-            base = region.base_address + self._delta
-            patched = self._apply_region_relocations(region)
-            placed.append((base, patched))
-        self._placed = placed
-        return placed
-
-    def _apply_region_relocations(self, region: Region) -> bytes:
-        if not region.expression_relocations:
-            return region.code
-        code_array = bytearray(region.code)
-        root_scope = self.resolver.scopes[0]
-        saved = self._inject_locals(root_scope)
-        try:
-            for offset, expr, size in region.expression_relocations:
-                self._eval_one_relocation(expr, offset, size, code_array, region)
-        finally:
-            self._restore_locals(root_scope, saved)
-        return bytes(code_array)
-
-    def _reloc_context(self, offset: int, region: Region) -> str:
-        return (
-            f"module '{self.module_name}' region@0x{region.base_address:x} offset 0x{offset:x}/0x{len(region.code):x}"
-        )
-
-    def _write_reloc(
-        self, code_array: bytearray, offset: int, value: int, size: int, expr: str, region: Region
-    ) -> None:
-        ctx = self._reloc_context(offset, region)
-        if size not in (1, 2, 3):
-            logger.warning(f"Unsupported relocation size {size} for expression '{expr}' [{ctx}]")
-            return
-        if offset + size > len(code_array):
-            logger.warning(
-                f"Relocation runs past region code: offset 0x{offset:x} + size {size} "
-                f"> 0x{len(code_array):x} for expression '{expr}' [{ctx}]"
-            )
-            return
-        for i in range(size):
-            code_array[offset + i] = (value >> (8 * i)) & 0xFF
-
-    def _eval_one_relocation(self, expr: str, offset: int, size: int, code_array: bytearray, region: Region) -> None:
-        from a816.parse.ast.expression import eval_expression_str
-
-        ctx = self._reloc_context(offset, region)
-        try:
-            value = eval_expression_str(expr, self.resolver)
-        except (SymbolNotDefined, NodeError, ValueError) as e:
-            logger.warning(f"Failed to evaluate expression '{expr}': {e} [{ctx}]")
-            return
-        if not isinstance(value, int):
-            logger.warning(f"Expression '{expr}' did not evaluate to int: {value} [{ctx}]")
-            return
-        self._write_reloc(code_array, offset, value, size, expr, region)
-
-    def _inject_locals(self, root_scope: Scope) -> dict[str, int | str]:
-        local_map = getattr(self, "_local_map", {})
-        saved: dict[str, int | str] = {}
-        for name, value in local_map.items():
-            if name in root_scope.symbols:
-                saved[name] = root_scope.symbols[name]
-            root_scope.symbols[name] = value
-        return saved
-
-    def _restore_locals(self, root_scope: Scope, saved: dict[str, int | str]) -> None:
-        local_map = getattr(self, "_local_map", {})
-        for name in local_map:
-            if name in saved:
-                root_scope.symbols[name] = saved[name]
-            else:
-                root_scope.symbols.pop(name, None)
-
-    def __str__(self) -> str:
-        total = sum(len(r.code) for r in self.regions)
-        return (
-            f"LinkedModuleNode({self.module_name}, {len(self.regions)} regions, "
-            f"{total} bytes, {len(self.symbols)} symbols)"
-        )
-
-
-class _SizedValueNode(NodeProtocol):
-    """Emit `SIZE` little-endian bytes from a value node.
-
-    Concrete subclasses set `SIZE` (1, 2, or 3). Deferred expressions
-    on the value node register an expression relocation of the same
-    width so the linker can fill it in once cross-module names resolve.
-    """
-
-    SIZE: int = 0
-
-    def __init__(self, value_node: ValueNodeProtocol) -> None:
-        self.value_node = value_node
-
-    def emit(self, current_address: Address) -> bytes:
-        value = self.value_node.get_value()
-        if isinstance(self.value_node, ExpressionNode) and hasattr(self.value_node, "_deferred_expression"):
-            resolver = self.value_node.resolver
-            if resolver.context.is_object_mode and resolver.context.object_writer is not None:
-                resolver.context.object_writer.add_expression_relocation(
-                    resolver.context.object_writer.relocation_offset(),
-                    self.value_node._deferred_expression,
-                    self.SIZE,
-                )
-        mask = (1 << (8 * self.SIZE)) - 1
-        return (value & mask).to_bytes(self.SIZE, "little")
-
-    def pc_after(self, current_pc: Address) -> Address:
-        return current_pc + self.SIZE
-
-
-class LongNode(_SizedValueNode):
-    SIZE = 3
-
-
-class WordNode(_SizedValueNode):
-    SIZE = 2
-
-
-class DebugNode(NodeProtocol):
-    def __init__(self, message: str, resolver: Resolver) -> None:
-        self.message = message
-        self.resolver = resolver
-
-    def emit(self, current_address: Address) -> bytes:
-        message = self.message
-        matches = re.finditer(r"\{([^}]+)}", self.message)
-
-        for match in matches:
-            expression_str = match.group(1)
-
-            _value = eval_expression_str(expression_str, self.resolver)
-
-            match _value:
-                case int():
-                    value = hex(_value)
-                case str():
-                    value = _value
-
-            message = message.replace(match.group(0), value)
-
-        print(message)
-        return b""
-
-    def pc_after(self, current_pc: Address) -> Address:
-        return current_pc
-
-
-class ByteNode(_SizedValueNode):
-    SIZE = 1
-
-
-class UnknownOpcodeError(Exception):
-    pass
-
-
-class NodeError(A816Error):
-    def __init__(
-        self,
-        message: str,
-        file_info: Token,
-        *,
-        code: str | None = None,
-        hint: str | None = None,
-    ) -> None:
-        super().__init__(message)
-        self.file_info = file_info
-        self.message = message
-        self.code = code
-        self.hint = hint
-
-    def __str__(self) -> str:
-        return self.format()
-
-    def format(self) -> str:
-        """Format the error with source location, visual indicator, code, and hint."""
-        # Late import: intentional to avoid circular dependency with errors module
-        from a816.errors import SourceLocation, format_error
-
-        location = None
-        if self.file_info is not None and self.file_info.position is not None:
-            pos = self.file_info.position
-            try:
-                source_line = pos.get_line()
-            except (IndexError, AttributeError):
-                source_line = ""
-            file = getattr(pos, "file", None)
-            lines = getattr(file, "lines", None)
-            context_before: list[str] | None = None
-            context_after: list[str] | None = None
-            if lines:
-                line_idx = pos.line
-                context_before = [lines[i] for i in range(max(0, line_idx - 1), line_idx)] or None
-                context_after = [lines[i] for i in range(line_idx + 1, min(len(lines), line_idx + 2))] or None
-            location = SourceLocation(
-                filename=pos.file.filename,
-                line=pos.line,
-                column=pos.column,
-                source_line=source_line,
-                length=len(self.file_info.value) if self.file_info.value else 1,
-                context_before=context_before,
-                context_after=context_after,
-            )
-
-        return format_error(self.message, location, code=self.code, hint=self.hint)
-
-
-class OpcodeNode(NodeProtocol):
-    def __init__(
-        self,
-        opcode: str,
-        *,
-        size: ValueSize | None = None,
-        addressing_mode: AddressingMode,
-        index: str | None = None,
-        value_node: ValueNodeProtocol | None = None,
-        file_info: Token,
-        resolver: Resolver,
-    ) -> None:
-        self.opcode = opcode.lower()
-        self.addressing_mode = addressing_mode
-        self.index = index
-        self.value_node = value_node
-        self.size = size
-        self.file_info = file_info
-        self.resolver = resolver
-
-    def _get_emitter(self) -> OpcodeProtocol:
-        try:
-            opcode_emitter = snes_opcode_table[self.opcode][self.addressing_mode]
-        except KeyError as e:
-            raise NodeError(
-                f"Addressing mode ({self.addressing_mode.name}) for opcode_def ({self.opcode}) is not defined.",
-                file_info=self.file_info,
-            ) from e
-
-        if isinstance(opcode_emitter, dict):
-            if self.index is not None:
-                opcode_emitter = opcode_emitter[self.index]
-            else:
-                raise NodeError(
-                    f"Addressing mode ({self.addressing_mode.name}) for opcode_def ({self.opcode}) needs an index.",
-                    file_info=self.file_info,
-                )
-        return opcode_emitter
-
-    def emit(self, current_pc: Address) -> bytes:
-        opcode_emitter = self._get_emitter()
-        try:
-            return opcode_emitter.emit(self.value_node, self.resolver, self.size)
-        except NoOpcodeForOperandSize as e:
-            assert self.value_node is not None
-            guessed_size = guess_value_size(self.value_node, self.size)
-            raise NodeError(
-                f"{self.opcode} does not supports size ({guessed_size}).",
-                self.file_info,
-            ) from e
-        except SymbolNotDefined as e:
-            raise NodeError(
-                f"`{e}` is not defined in the current scope",
-                self.file_info,
-                code=str(_E_SYMBOL_NOT_DEFINED),
-                hint=_did_you_mean_hint(str(e), self.resolver.current_scope),
-            ) from e
-
-    def pc_after(self, current_pc: Address) -> Address:
-        opcode_emitter = self._get_emitter()
-        return current_pc + opcode_emitter.supposed_length(self.value_node, self.size)
-
-    def __str__(self) -> str:
-        return f"OpcodeNode({self.opcode}, {self.addressing_mode}, {self.index}, {self.value_node})"
-
-
-class AllocNode(NodeProtocol):
-    """Emits `body` at an address picked by the named pool's allocator.
-
-    `pc_after` runs once per resolver pass; the pool allocator is invoked
-    by `Resolver.allocate_pools()` between passes. The body is walked
-    first to measure its size, the slot is requested, and once placed the
-    `name` label binds at the allocated address.
-    """
-
-    def __init__(
-        self,
-        name: str,
-        pool_name: str,
-        body: list[NodeProtocol],
-        resolver: "Resolver",
-        file_info: Token,
-    ) -> None:
-        self.name = name
-        self.pool_name = pool_name
-        self.body = body
-        self.resolver = resolver
-        self.file_info = file_info
-        self._alloc: Allocation | None = None
-        self._size: int = 0
-
-    def _sandbox_pc(self) -> Address:
-        pool = self.resolver.pools[self.pool_name]
-        anchor = pool.ranges[0].start if pool.ranges else 0
-        return self.resolver.get_bus().get_address(anchor)
-
-    @staticmethod
-    def _skip_in_pass1(node: NodeProtocol) -> bool:
-        """Mirror `Program.resolve_labels` pass-1 skip set.
-
-        Nodes whose `pc_after` eagerly evaluates expressions (e.g.
-        `SymbolNode`) must wait for pass-2 or forward references raise.
-        AllocNode walks its body across the same passes, so applies the
-        same skip rule when measuring + when binding labels.
-        """
-        return isinstance(node, SymbolNode)
-
-    def _measure_body(self) -> int:
-        start = self._sandbox_pc()
-        pc = start
-        for node in self.body:
-            if self._skip_in_pass1(node):
-                continue
-            pc = node.pc_after(pc)
-        return pc.logical_value - start.logical_value
-
-    def _request_slot(self) -> None:
-        pool = self.resolver.pools[self.pool_name]
-        self._size = max(1, self._measure_body())
-        self._alloc = pool.request(self.name, self._size)
-        # Object mode defers allocator to link time. Bind the alloc's
-        # symbol + body labels at the sandbox PC (pool.ranges[0].start)
-        # so they record sensible offsets; the linker rebases the body
-        # region at link time and the existing CODE-symbol delta path
-        # carries every label to its final address.
-        if self.resolver.context.is_object_mode:
-            self._bind_body_labels_at(self._sandbox_pc())
-
-    def _bind_body_labels_at(self, target: Address) -> None:
-        self.resolver.current_scope.add_label(self.name, target)
-        pc = target
-        for node in self.body:
-            pc = node.pc_after(pc)
-
-    def _bind_body_labels(self) -> None:
-        alloc = self._alloc
-        assert alloc is not None
-        target = self.resolver.get_bus().get_address(alloc.addr)
-        self._bind_body_labels_at(target)
-
-    def pc_after(self, current_pc: Address) -> Address:  # NOSONAR S3516
-        # Returning current_pc unchanged is by design: an .alloc block emits
-        # at the allocator-chosen address (via emit_blocks), not inline,
-        # so the surrounding PC must not advance past this node.
-        pool = self.resolver.pools.get(self.pool_name)
-        if pool is None:
-            raise NodeError(f".alloc into unknown pool {self.pool_name!r}", self.file_info)
-        if self._alloc is None:
-            self._request_slot()
-        elif self._alloc.placed:
-            self._bind_body_labels()
-        return current_pc
-
-    def emit(self, current_addr: Address) -> bytes:
-        del current_addr
-        return b""
-
-    def emit_blocks(self, current_addr: Address) -> list[tuple[int, bytes]]:
-        del current_addr
-        alloc = self._alloc
-        if alloc is None or not alloc.placed:
-            return []
-        # Lazy import to avoid the import cycle through codegen.
-        from a816.parse.nodes import LinkedModuleNode
-
-        saved_pc = self.resolver.pc
-        saved_reloc = self.resolver.reloc_address
-        try:
-            self.resolver.set_position(alloc.addr)
-            out = b""
-            cur = self.resolver.reloc_address
-            for node in self.body:
-                # Skip loser `.import` duplicates inside the body — their
-                # pc_after returned unchanged so subsequent siblings are
-                # already laid out at the right offsets, and emitting the
-                # loser bytes would both double-emit and overlap the
-                # winner's placement. Mirrors the top-level
-                # `_emit_linked_module` is_loser guard.
-                if isinstance(node, LinkedModuleNode) and node.is_loser:
-                    continue
-                emitted = node.emit(cur)
-                if emitted:
-                    out += emitted
-                    cur = cur + len(emitted)
-                    self.resolver.pc += len(emitted)
-                    self.resolver.reloc_address = cur
-            return [(alloc.addr, out)]
-        finally:
-            self.resolver.pc = saved_pc
-            self.resolver.reloc_address = saved_reloc
-
-    def __str__(self) -> str:
-        return f"AllocNode({self.name} in {self.pool_name})"
-
-
-class RelocateNode(AllocNode):
-    """`AllocNode` that also reclaims the symbol's original (addr, end) range.
-
-    Pass 1 reclaims `[old_start, old_end]` back into the pool *before*
-    requesting a slot for the body, so the reclaimed bytes may be reused
-    by the same `.relocate` if the pool's free space is otherwise full.
-    """
-
-    def __init__(
-        self,
-        name: str,
-        old_start: int,
-        old_end: int,
-        pool_name: str,
-        body: list[NodeProtocol],
-        resolver: "Resolver",
-        file_info: Token,
-    ) -> None:
-        super().__init__(name, pool_name, body, resolver, file_info)
-        self.old_start = old_start
-        self.old_end = old_end
-        self._reclaimed = False
-
-    def _request_slot(self) -> None:
-        if not self._reclaimed:
-            from a816.pool import PoolRange
-
-            pool = self.resolver.pools[self.pool_name]
-            try:
-                pool.reclaim(PoolRange(start=self.old_start, end=self.old_end))
-            except Exception as exc:
-                raise NodeError(
-                    f".relocate {self.name!r} reclaim 0x{self.old_start:06x}..0x{self.old_end:06x}: {exc}",
-                    self.file_info,
-                ) from exc
-            self._reclaimed = True
-        super()._request_slot()
-
-    def __str__(self) -> str:
-        return f"RelocateNode({self.name} from 0x{self.old_start:06x}..0x{self.old_end:06x} -> {self.pool_name})"
-
-
-class CodePositionNode(NodeProtocol):
-    def __init__(self, value_node: ValueNodeProtocol, resolver: Resolver):
-        self.value_node = value_node
-        self.resolver: Resolver = resolver
-
-    def pc_after(self, current_pc: Address) -> Address:
-        self.resolver.reloc = False
-        return self.resolver.get_bus().get_address(self.value_node.get_value())
-
-    def emit(self, current_addr: Address) -> bytes:
-        self.resolver.set_position(self.value_node.get_value())
-        return b""
-
-    def __str__(self) -> str:
-        return f"CodePositionNode({self.value_node.get_value()})"
-
-
-class RelocationAddressNode(NodeProtocol):
-    def __init__(self, pc_value_node: ValueNodeProtocol, resolver: Resolver) -> None:
-        self.pc_value_node = pc_value_node
-        self.resolver = resolver
-
-    def pc_after(self, current_pc: Address) -> Address:
-        self.resolver.reloc = True
-        return self.resolver.get_bus().get_address(self.pc_value_node.get_value())
-
-    def emit(self, current_addr: Address) -> bytes:
-        self.resolver.set_position(self.pc_value_node.get_value())
-        # self.resolver.set_position(self.pc_value_node.get_value(), reloc=True)
-        return b""
-
-    def __str__(self) -> str:
-        return f"RelocationAddressNode({self.pc_value_node.get_value()})"
-
-
-class IncludeIpsNode(NodeProtocol):
-    def __init__(
-        self,
-        file_path: str,
-        resolver: Resolver,
-        delta_expression: ExpressionAstNode | None = None,
-    ) -> None:
-        self.ips_file_path = file_path
-        self.delta = eval_expression(delta_expression, resolver) if delta_expression else 0
-        self.blocks: list[tuple[int, bytes]] = []
-        with open(self.ips_file_path, "rb") as ips_file:
-            if ips_file.read(5) != b"PATCH":
-                raise RuntimeError(f'{self.ips_file_path} is missing "PATCH" header')
-
-            while ips_file.peek(3)[:3] != b"EOF":
-                block_addr_bytes = struct.unpack(">BH", ips_file.read(3))
-                block_addr = (block_addr_bytes[0] << 16) | block_addr_bytes[1]
-                block_size_word = struct.unpack(">H", ips_file.read(2))
-                block_size = block_size_word[0]
-                block = ips_file.read(block_size)
-
-                if self.delta is not None:
-                    block_addr += self.delta
-
-                self.blocks.append((block_addr, block))
-
-    def pc_after(self, current_pc: Address) -> Address:
-        return current_pc
-
-    def emit(self, current_addr: Address) -> bytes:
-        return b""
-
-
-class ScopeNode(NodeProtocol):
-    def __init__(self, resolver: Resolver) -> None:
-        self.resolver = resolver
-        self.parent_scope = self.resolver.current_scope
-
-    def pc_after(self, current_pc: Address) -> Address:
-        self.resolver.use_next_scope()
-        return current_pc
-
-    def emit(self, current_addr: Address) -> bytes:
-        self.resolver.use_next_scope()
-        return b""
-
-
-class PopScopeNode(NodeProtocol):
-    def __init__(self, resolver: Resolver) -> None:
-        self.resolver = resolver
-
-    def pc_after(self, current_pc: Address) -> Address:
-        self.resolver.restore_scope(exports=True)
-        return current_pc
-
-    def emit(self, current_addr: Address) -> bytes:
-        self.resolver.restore_scope()
-        return b""
-
-
-class TableNode(NodeProtocol):
-    def __init__(self, path: str, resolver: Resolver) -> None:
-        self.table_path = path
-        self.resolver = resolver
-        resolver.current_scope.table = Table(self.table_path)
-
-    def pc_after(self, current_pc: Address) -> Address:
-        return current_pc
-
-    def emit(self, current_addr: Address) -> bytes:
-        return b""
-
-
-class AbstractTextNode(NodeProtocol):
-    def __init__(self, text: str, resolver: Resolver) -> None:
-        self.text = text
-        self.resolver = resolver
-
-    @property
-    def binary_text(self) -> bytes:
-        raise NotImplementedError("You should implement binary_text property.")
-
-    def pc_after(self, current_pc: Address) -> Address:
-        return current_pc + len(self.binary_text)
-
-    def emit(self, current_addr: Address) -> bytes:
-        return self.binary_text
-
-
-def variable_expansion(value: str, resolver: Resolver) -> str:
-    if value is not None and isinstance(value, str):
-
-        def replace_match(match: re.Match[str]) -> str:
-            lookup = match.groups("lookup")[0]
-            variable_value = resolver.current_scope.value_for(lookup)
-            if variable_value is None:
-                raise ValueError(f"Error while resolving variable {lookup} in {value} variable value is None.")
-            # Convert to string for substitution
-            if isinstance(variable_value, int):
-                return str(variable_value)
-            elif isinstance(variable_value, str):
-                return variable_value
-            else:
-                # BlockAstNode - convert to string representation
-                return str(variable_value)
-
-        try:
-            return re.sub(r"\${(?P<lookup>[^}]+)}", replace_match, value) or value
-        except ValueError:
-            return value
-
-
-class TextNode(AbstractTextNode):
-    def __init__(self, text: str, resolver: Resolver, file_info: Token) -> None:
-        super().__init__(text, resolver)
-        self.table = self.resolver.current_scope.get_table()
-        self.file_info = file_info
-
-    @property
-    def binary_text(self) -> bytes:
-        if self.table is None:
-            raise NodeError(
-                f"table_is_not_defined ({self}) is not defined in the current scope.",
-                self.file_info,
-            )
-        return self.table.to_bytes(variable_expansion(self.text, self.resolver))
-
-
-class AsciiNode(AbstractTextNode):
-    @property
-    def binary_text(self) -> bytes:
-        return self.text.encode("ascii", errors="ignore")
-
-
-class PointerNode(NodeProtocol):
-    def __init__(self, value_node: ValueNodeProtocol) -> None:
-        self.value_node = value_node
-
-    def pc_after(self, current_pc: Address) -> Address:
-        return current_pc + 3
-
-    def emit(self, current_addr: Address) -> bytes:
-        value = self.value_node.get_value()
-        return struct.pack("<HB", value & 0xFFFF, (value >> 16) & 0xFF)
